Pass actions for found/not-found to AX_CHECK_GLUT.
AX_CHECK_GLUT from autoconf-archive 20160320 gives an error
when GLUT is not found and no action for not-found is set.

Set no_glut=yes for compatibility with the earlier version.

Fixes this error when SDL is installed but GLUT is not:

configure: error: Could not find a valid GLUT implementation